kda interview: 修复 triton-ascend 3.2.1 / CANN 9.1 工具链适配问题
- K2: tl.gather(src=tl.dot 输出) 在 3.2.1 上数值错误(Akk max_diff≈0.32)且退化
  ~3x 慢，driver 切回 hm2 路径(满宽写 scratch + torch.gather, ≈8.9e-8)；
  hm3 kernel 标记 DEPRECATED 仅作参考
- K3: Akk 系数 beta_j 修正为 beta_i；D_inv 符号修正为 (I + tril(D))^{-1}
- K5: store 统一走 _store_h_full 通用 2D 指针(flat reshape store 触发 MLIR
  expand_shape 报错)；外积改为 b_v^T @ k，规避 hivm-plan-memory 误报 ub overflow
- K6: 因果 mask 改用 bool(fp32 compare+where 在 3.2.1 误编译)
- PROBLEM.md ×6: 补设计解析与 09-07 msprof 实测基线

+++ kda_test/interview/k6_gla_output/gla_output_kernel.py
@@ -289,7 +289,7 @@ def chunk_gla_fwd_kernel_o(
     s_a_t: tl.constexpr = H * BT      # A 在 T 维的 stride
 
     # 下三角因果 mask: m_s[i, j] = (i >= j)  (fp32)
-    m_s = tl.arange(0, BT)[:, None].to(tl.float32) >= tl.arange(0, BT)[None, :].to(tl.float32)
+    m_s = tl.arange(0, BT)[:, None] >= tl.arange(0, BT)[None, :]   # bool 掩码（勿改 fp32: 3.2.1 对 fp32 where→dot 输入误编译）
 
     b_o = tl.zeros([BT, BV], dtype=tl.float32)
 
@@ -374,7 +374,7 @@ def chunk_gla_fwd_kernel_o_hm(
     # ── 循环不变向量（与 head 无关，全部提升到头循环外）──
     r = tl.arange(0, BT)
     c = tl.arange(0, BT)
-    m_s = r[:, None].to(tl.float32) >= c[None, :].to(tl.float32)   # 因果 mask [BT,BT]
+    m_s = r[:, None] >= c[None, :]   # 因果 mask [BT,BT] (bool；勿改 fp32 compare + where → 3.2.1 误编译见 OPTIMIZATION_LOG)
     r_mask = (i_t * BT + r) < T
     k_mask = tl.arange(0, BK) < K
     v_mask = (i_v * BV + tl.arange(0, BV)) < V
